Change len argument of string RsValues to u32 everywhere

+++ src/redisearch_rs/c_entrypoint/value_ffi/src/setters.rs
@@ -8,7 +8,6 @@
 */
 
 use crate::util::expect_shared_value;
-use libc::size_t;
 use std::ffi::{c_char, c_double};
 use value::{RsString, RsValue};
 
@@ -77,12 +76,12 @@ pub unsafe extern "C" fn RSValue_SetNull(value: *mut RsValue) {
 ///
 /// Panics if more than 1 reference exists to this [`RsValue`] object.
 #[unsafe(no_mangle)]
-pub unsafe extern "C" fn RSValue_SetString(value: *mut RsValue, str: *mut c_char, len: size_t) {
+pub unsafe extern "C" fn RSValue_SetString(value: *mut RsValue, str: *mut c_char, len: u32) {
     // Safety: ensured by caller (1.)
     let mut shared_value = unsafe { expect_shared_value(value) };
 
     // Safety: ensured by caller (2., 3., 4.)
-    let string = unsafe { RsString::rm_alloc_string(str, len as u32) };
+    let string = unsafe { RsString::rm_alloc_string(str, len) };
     let value = RsValue::String(string);
     // Safety: ensured by caller (5.)
     shared_value.set_value(value);
@@ -109,16 +108,12 @@ pub unsafe extern "C" fn RSValue_SetString(value: *mut RsValue, str: *mut c_char
 ///
 /// Panics if more than 1 reference exists to this [`RsValue`] object.
 #[unsafe(no_mangle)]
-pub unsafe extern "C" fn RSValue_SetConstString(
-    value: *mut RsValue,
-    str: *const c_char,
-    len: size_t,
-) {
+pub unsafe extern "C" fn RSValue_SetConstString(value: *mut RsValue, str: *const c_char, len: u32) {
     // Safety: ensured by caller (1.)
     let mut shared_value = unsafe { expect_shared_value(value) };
 
     // Safety: ensured by caller (2., 3., 4.)
-    let string = unsafe { RsString::borrowed_string(str, len as u32) };
+    let string = unsafe { RsString::borrowed_string(str, len) };
     let value = RsValue::String(string);
     // Safety: ensured by caller (5.)
     shared_value.set_value(value);
